pyproject.toml
[tool.poetry]
name = "smith"
version = "0.1.0"
description = ""
authors = ["Juan Martínez <dahbura.ml@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"
langchain-core = "^0.1.36"
langchain-anthropic = "^0.1.4"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
